+++ src/commands/UploadConnectorCommand.ts
@@ -1,13 +1,10 @@
 import * as vscode from "vscode";
-import * as fs from 'fs';
-import * as path from 'path';
 
 import { ConnectorTreeItem } from "../models/TreeModel";
 import { SaaSConnectivityClientFactory } from "../services/SaaSConnectivityClientFactory";
 import { ISCExtensionClient } from "../iscextension/iscextension-client";
-import { chooseFileExtended, confirm, getWorkspaceFolder } from "../utils/vsCodeHelpers";
+import { chooseFileExtended, confirm } from "../utils/vsCodeHelpers";
 import { UploadConnectorResponse } from "../models/API";
-import { isEmpty } from "../utils/stringUtils";
 
 
 
@@ -75,61 +72,3 @@ export class UploadConnectorCommand extends ConnectorUploader {
     }
 }
 
-/**
- * Will try to automatically find the zip file
- */
-export class DeployConnectorCommand extends ConnectorUploader {
-    constructor() {
-        super()
-    }
-
-    async chooseFile(): Promise<undefined | string> {
-        const workspaceFolder = getWorkspaceFolder()
-        console.log({ workspaceFolder });
-
-        if (workspaceFolder === undefined) {
-            vscode.window.showErrorMessage("No workspace found. Try to open the folder of your connector before deploying.")
-            return undefined
-        }
-
-        const packagePath = path.join(workspaceFolder, "package.json")
-        console.log({ packagePath });
-
-        let packageContent: string
-        try {
-            packageContent = fs.readFileSync(packagePath, { encoding: "utf8" })
-        } catch (error) {
-            console.error(error);
-            vscode.window.showErrorMessage("Could not read `package.json`. Try to open the folder of your connector before deploying.")
-            return undefined
-        }
-
-        let version: string
-        let name: string
-        try {
-            const packageJSON = JSON.parse(packageContent)
-            version = packageJSON.version
-            name = packageJSON.name
-            if (isEmpty(version) || isEmpty(name)) {
-                throw new Error("Version or name empty.");
-            }
-        } catch (error) {
-            console.error(error);
-            vscode.window.showErrorMessage("Invalid `package.json`. Try to open the folder of your connector before deploying.")
-            return undefined
-        }
-
-        const zipPath = path.join(workspaceFolder, "dist", `${name}-${version}.zip`)
-        console.log({ zipPath });
-
-        fs.access(zipPath, fs.constants.R_OK, (err) => {
-            if (err) {
-                console.error(err);
-                vscode.window.showErrorMessage(`Could not find ${zipPath}. Run \`npm run pack-zip\` to build and package the connector bundle`)
-                return undefined
-            }
-        })
-        return zipPath
-
-    }
-}